`fn dav1d_decode_frame_main`: Make `tile_{row,col}` `usize`s.

+++ src/decode.rs
@@ -5752,27 +5752,28 @@ pub unsafe extern "C" fn dav1d_decode_frame_main(f: *mut Dav1dFrameContext) -> l
 
     // no threading - we explicitly interleave tile/sbrow decoding
     // and post-filtering, so that the full process runs in-line
-    for tile_row in 0..(*f.frame_hdr).tiling.rows {
+    for tile_row in 0..(*f.frame_hdr).tiling.rows as usize {
         let sbh_end = std::cmp::min(
-            (*f.frame_hdr).tiling.row_start_sb[(tile_row + 1) as usize].into(),
+            (*f.frame_hdr).tiling.row_start_sb[tile_row + 1].into(),
             f.sbh,
         );
-        for sby in (*f.frame_hdr).tiling.row_start_sb[tile_row as usize].into()..sbh_end {
+        for sby in (*f.frame_hdr).tiling.row_start_sb[tile_row].into()..sbh_end {
             t.by = sby << 4 + (*f.seq_hdr).sb128;
             let by_end = t.by + f.sb_step >> 1;
             if (*f.frame_hdr).use_ref_frame_mvs != 0 {
                 ((*f.c).refmvs_dsp.load_tmvs).expect("non-null function pointer")(
                     &mut f.rf,
-                    tile_row,
+                    tile_row as libc::c_int,
                     0,
                     f.bw >> 1,
                     t.by >> 1,
                     by_end,
                 );
             }
-            for tile_col in 0..(*f.frame_hdr).tiling.cols {
-                t.ts =
-                    f.ts.offset((tile_row * (*f.frame_hdr).tiling.cols + tile_col) as isize);
+            for tile_col in 0..(*f.frame_hdr).tiling.cols as usize {
+                t.ts = f
+                    .ts
+                    .offset((tile_row * (*f.frame_hdr).tiling.cols as usize + tile_col) as isize);
                 if dav1d_decode_tile_sbrow(t) != 0 {
                     return retval;
                 }
